export interface from frontend and import them globally in backend controllers files
