+++ client/src/models/Ingredient.ts
@@ -1,0 +1,7 @@
+export interface IIngredient extends Document {
+    _id: string;
+    name: string;
+    unitCost: number;
+    quantity: number;
+    thresholdLevel: number;
+}
