+++ server/controllers/menu.ts
@@ -1,24 +1,10 @@
 import express from "express";
 import { Collection } from "mongodb";
 import { Client_Connect } from "../../config/config.ts";
-
-interface IMenuIngredient {
-  ingredientName: string;
-  ingredientId: string;
-}
+import {IMenu} from "../../client/src/models/Menu.ts";
 
 const router = express.Router();
 
-interface IMenu extends Document {
-  _id: string;
-  name: string;
-  category: string;
-  ingredients: IMenuIngredient;
-  quantity: number;
-  price: number;
-  prepTime: number;
-  Image: string;
-}
 router.get("/menu", async (req, res) => {
   try {
     // init db connection with MongoClient
